change getqvars parameters from int * to unsigned long *

+++ lab11/kernelsrc/stoogecalls/getqvars.c
@@ -3,7 +3,7 @@
 #include <linux/syscalls.h>
 #include <linux/blkdev.h>
 
-SYSCALL_DEFINE3(getqvars, int *, service_time, int *, wait_time, int *, num_requests)
+SYSCALL_DEFINE3(getqvars, unsigned long *, service_time, unsigned long *, wait_time, unsigned long *, num_requests)
 {
     *service_time = q_total_service_time;
     *wait_time = q_total_wait_time;
